requirements.txt
blinker==1.9.0
click==8.2.1
colorama==0.4.6
filelock==3.13.1
Flask==3.1.2
flask-cors==6.0.1
fsspec==2024.6.1
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.3
numpy==2.2.6
opencv-python==4.12.0.88
pillow==11.3.0
setuptools==70.2.0
sympy==1.13.3
torch==2.8.0+cu128
torchvision==0.23.0+cu128
typing_extensions==4.12.2
Werkzeug==3.1.3
PyYAML==6.0.1
fastapi==0.116.1
uvicorn==0.35.0
